requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
geoalchemy2
pydantic
requests
paho-mqtt